Handling of omitted mortality variable
The mortality input variable is necessary to generate the table used to train
the model, but not necessary when making predictions on new input data. This
commit includes code to scan the input file for the inclusion or omission of
of the mortality variable, and modify the output table accordingly.